pyproject.toml
[tool.poetry]
name = "my_pipeline"
version = "0.1.0"
description = "A genomic variant calling pipeline using Snakemake and ML filtering."
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/yourusername/my_pipeline"
homepage = "https://github.com/yourusername/my_pipeline"
keywords = ["genomics", "variant-calling", "ml", "snakemake", "bioinformatics"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Topic :: Scientific/Engineering :: Bio-Informatics"
]

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
pysam = "^0.21.0"
numpy = "^1.20.0"
PyYAML = "^6.0"
scikit-learn = "^1.2.0"
snakemake = {version = "^7.0.0", optional = true}

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"

[tool.poetry.scripts]
# Example: Make `my_pipeline-run` a CLI command to run a specific script
my_pipeline-run = "my_pipeline.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
